rewrote it all again
Ok, so turns out, we need to more drastically rewrite every protobuf so
we don't clobber and cause conflicts in the global registry. This means
package names truly need to be unique.

So this rewrites everything into a `vitess.{pkg}.{version]` format while
also rewriting every reference and every import.