Hack to fix service name generation
This is needed since we're repackaging the vitess protos into unique
packages, but then they fail to interop with a vitess server, so we need
to re-format and strip the versions out of the packages.

This is fine since we need the boundary to stay the same and Vitess
itself doesn't strongly version their APIs, so we need to make sure we
call the correct URLs.